+++ packages/navigation/src/Menu/index.style.tsx
@@ -4,6 +4,7 @@ import { bg } from '@lendi-ui/color';
 import { depth } from '@lendi-ui/depth';
 import { ChevronUp, ChevronDown } from '@lendi-ui/icon';
 import { deriveSize } from '@lendi-ui/utils';
+import { select } from '@lendi-ui/theme';
 
 export const MenuWrapper = styled.div`
     width: 100%;
@@ -11,7 +12,7 @@ export const MenuWrapper = styled.div`
     flex-flow: column;
     justify-content: center;
     align-items: center;
-    border-radius: 6px;
+    border-radius: ${select('borderRadius')};
     ${depth(1)} ${my('xs')}
 `;
 
@@ -28,11 +29,14 @@ export const HeadWrapper = styled.button`
     cursor: pointer;
     ${pl('xxs')} ${pr('xs')} ${py('xs')}
     ${bg('secondary.500')}
-    ${({ isOpen }: { isOpen: boolean }) => (isOpen ? 'border-radius: 6px 6px 0 0;' : 'border-radius: 6px;')}
+    ${({ isOpen }: { isOpen: boolean }) =>
+      isOpen
+        ? `border-radius: ${select('borderRadius')} ${select('borderRadius')} 0 0;`
+        : `border-radius: ${select('borderRadius')};`}
 `;
 
 export const MainWrapper = styled.div`
-  border-radius: 0 0 6px 6px;
+  border-radius: 0 0 ${select('borderRadius')} ${select('borderRadius')};
   ${bg('shade.0')} ${py('xxxs')};
 `;