+++ packages/card/CHANGELOG.md
@@ -2,8 +2,11 @@
 ### Changed
 - Updated CHANGELOG to match required parsing format.
 
-## [3.0.2] - 2019-05-08 - [HUB-54] (https://creditandfinance.atlassian.net/browse/HUB-54)
- 
+## [3.0.3] - 2019-05-02 - [HUB-88](https://creditandfinance.atlassian.net/browse/HUB-88)
+### Changed
+- replace `border-radius: 6px` with select function in theme
+
+## [3.0.2] - 2019-05-08 - [HUB-54](https://creditandfinance.atlassian.net/browse/HUB-54)
 ### Changed
 - Change module main entry point to esm in package.json