+++ packages/text-bubble/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [2.0.4] - 2019-05-02 - [HUB-88](https://creditandfinance.atlassian.net/browse/HUB-88)
+### Changed
+- replace `border-radius: 6px` with select function in theme
+
 ## [2.0.3] - 2019-05-09 - [HUB-92](https://creditandfinance.atlassian.net/browse/HUB-92)
 ### Changed
 - Updated CHANGELOG to match required parsing format.